+++ include/openpose/wrapper/wrapperAuxiliary.hpp
@@ -20,14 +20,14 @@ namespace op
      * @param userOutputWsEmpty
      * @param threadManagerMode
      */
-	OP_API void wrapperConfigureSecurityChecks(const WrapperStructPose& wrapperStructPose,
-                                             const WrapperStructFace& wrapperStructFace,
-                                             const WrapperStructHand& wrapperStructHand,
-                                             const WrapperStructInput& wrapperStructInput,
-                                             const WrapperStructOutput& wrapperStructOutput,
-                                             const bool renderOutput,
-                                             const bool userOutputWsEmpty,
-                                             const ThreadManagerMode threadManagerMode);
+    OP_API void wrapperConfigureSecurityChecks(const WrapperStructPose& wrapperStructPose,
+                                               const WrapperStructFace& wrapperStructFace,
+                                               const WrapperStructHand& wrapperStructHand,
+                                               const WrapperStructInput& wrapperStructInput,
+                                               const WrapperStructOutput& wrapperStructOutput,
+                                               const bool renderOutput,
+                                               const bool userOutputWsEmpty,
+                                               const ThreadManagerMode threadManagerMode);
 }
 
 #endif // OPENPOSE_WRAPPER_WRAPPER_AUXILIARY_HPP