+++ doc/release_notes.md
@@ -195,6 +195,7 @@ OpenPose Library - Release Notes
     23. OpenPose small GUI shows the frame number w.r.t. the original producer, rather than the frame id. E.g., if video is started at frame 30, OpenPose will display 30 rather than 0 in the first frame.
     24. OpenPose GUI: 'l' and 'k' functionality swapped.
     25. 3-D reconstruction module: Added flag `--3d_min_views` to select minimum number of cameras required for 3-D reconstruction.
+    26. Flag `--camera_fps` also applies to recorded video (`--write_video`).
 2. Functions or parameters renamed:
     1. Flag `no_display` renamed as `display`, able to select between `NoDisplay`, `Display2D`, `Display3D`, and `DisplayAll`.
     2. 3-D reconstruction demo is now inside the OpenPose demo binary.